Adding validation back to customer
